Transfer learning approach to training end-to-end
Froze base model (single-visit reconstruction network) weights and appended
layers for multi-visit reconstruction network (unet_block) to compare end-to-end
and modular model. From this analysis it appears that the data consistency
block after the unet block is negatively impacting the results. Further investiagition
required to determine reason for this behavior as it should be improving the accuracy
of the multi-visit reconstruction.